requirements.txt
flask>=3.0
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
